More tidying
